API: move pim and areadetector out of the epics folder
