+++ pcdsdevices/areadetector/cam.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 """
 PCDS cams and overrides for ophyd Cams.
