+++ pcdsdevices/areadetector/plugins.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 """
 PCDS plugins and Overrides for AreaDetector Plugins.
 """
